plotted computation graphs for both resnet and unet
